refactor error management
It would seem the standard solution asks for using exceptions raised
for the edge cases and the conditions stormy and full_capacity.
Previous solution done without using errors, just returning nil when
one of those conditions appeared, because customer did not ask for
capturing the exact condition that made the airport not to land or
takeoff.  Anyway that was useful to practice the exceptions.