Created take_off method and two test for it.
Created new method, take_off. It delete a plane in base of the  index you give as argument.

Re organized the order of the tests in airport_class_spec.rb file and created two new tests.
One check if the class Airport respond to the take_off method.
Another one check if it really work, if it is remove the plane you choose from its hangar.

-Created second domain models in the Tables.md file.
-Named the object Airport.new and Plane.new in airport_class-spec.rb file.
-Rspec all green.
-Rubocop very happy.
-Coverage tests 100%