Created weather array, now take_off raise an error if weather.sample is 'stormy'
Created a new instance variable @weather which is an array that contain the strings
'sunny' and 'stormy' which rappresents the weather.
Calling take_off one of them is choosen, and if is 'stormy' the plane
do not fly, doesn't get deleted by from the @hangars array.

Created a test to check if the take_off method raise up an  error when
stormy is selected by sample.
Riorganized the tests divided in some moredescribe block.

-Created third table
-Added attr_accessor for @weather in order to be able to make the right tests(always stormy, or sunny)
-Rubocop Happy
-All tests green
-Coverage 100%

Further paragraphs come after blank lines.

 - Bullet points are okay, too

 - Typically a hyphen or asterisk is used for the bullet, preceded
   by a single space, with blank lines in between, but conventions
   vary here

If you use an issue tracker, put references to them at the bottom,
like this:

Resolves: #123
See also: #456, #789