Prevented landing if weather is stormy and land when airport full + tests
Now Airport class can accept 1 argument, which will be the alua of the instance
variable @capacity. Created new method 'full?' which return true if the number of Planes in the @hangar
are equal to the airport capacity.
Now land raise up an error if the airport is full and if the weather is stormy.

Riorganized a little bit the airport spec file, with new describe. But will be improved later.
Have been added new tests to check if Airport.new can take 1 argument, if land raise up an error
when the airport is full. Another test to check if lan() raise up an error when weather stormy.

- 2 new tables created.
- Coverage 100%
- Rubocop Happy
- All tests green