Add capacity to airport
Adds a default capacity to airport to proevent a plane landing if full. A new
`full?` method enables the airport controller to access if the airport is at
capacity.

Previously messages were being sent to the plane to land, but this step required
a refactor to send messages to the airport to land the plane.