Update Airport and Plane to use Airport capacity when planes land
The land method in the Plane class was updated to tell the airport
it is landing. This updates a member variable within the airport
instance that keeps track of how many planes are currently landed.

The full? method in the Airport class was updated to check against
the number of planes landed and correctly return whether it is full
or not.

The usage example in the README for user story 3 was updated to
reflect these changes.