Raise an error if a plane tries to take off in stormy weather
This was implemented by passing the weather into the take_off
method as an argument, which required changes to other uses of
that method across the test suite.

In addition, I also refactored the Weather class, introducing
constants to represent the possible weather conditions. These are
now used across the code base.