Raise an error if a plane tries to land in stormy weather
The land method in the Plane class now raises an error if a plane
tries to land when the weather is stormy. This is implemented by
passing in the current weather as an argument to the method, which
required changes to other areas of the code that call the same
method.