Refactor to move Weather method to a Class method
The Weather class was rewritten turning its only method from an
instance into a class method. It was also renamed from
current_weather to now. This means that you can query the class
for the current weather conditions with Weather.now, which is far
simpler syntax than before. It also means you no longer have to
pass an instance of Weather into the Plane class's land and
take_off methods.

This change led to significant alterations to much of the testing
code, and a rewrite of the README to reflect the changes in the
interfaces to the various classes.