User Story 4 - It can have a default capacity that can be overriden
It changes the tests expecations since the initialize method now stores the planes in an array.
Also, in the #land a plane is added to the planes array insted of a number being returned.
The #take_off removes and returns a plane from the planes array instead of subtracting 1 from the instance variable plane.
An attribute accessor capacity has been added.
An attribute reader has been added to help while feature testing in irb